requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.0
statsmodels==0.14.0
scipy==1.11.3
scikit-learn==1.3.2
seaborn==0.13.0
plotly==5.17.0
pywebview==4.3.1
openpyxl==3.1.2
xlrd==2.0.1